requirements.txt
opencv-python
numpy
keyboard
mediapipe
pillow
pystray
pyinstaller
